Extracts Remote::File
