+++ lib/pansophy/remote.rb
@@ -1,1 +1,2 @@
 require 'pansophy/remote/directory'
+require 'pansophy/remote/file'
